+++ python/cudf/cudf/core/column/string.py
@@ -86,6 +86,7 @@ from cudf._lib.strings.convert.convert_urls import (
 from cudf._lib.strings.extract import extract as cpp_extract
 from cudf._lib.strings.find import (
     contains as cpp_contains,
+    contains_multiple as cpp_contains_multiple,
     endswith as cpp_endswith,
     endswith_multiple as cpp_endswith_multiple,
     find as cpp_find,
@@ -620,12 +621,20 @@ class StringMethods(ColumnMethodsMixin):
         elif na is not np.nan:
             raise NotImplementedError("`na` parameter is not yet supported")
 
-        return self._return_or_inplace(
-            cpp_contains_re(self._column, pat)
-            if regex is True
-            else cpp_contains(self._column, as_scalar(pat, "str")),
-            **kwargs,
-        )
+        if pat is None:
+            result_col = column.column_empty(
+                len(self._column), dtype="bool", masked=True
+            )
+        elif is_scalar(pat):
+            if regex is True:
+                result_col = cpp_contains_re(self._column, pat)
+            else:
+                result_col = cpp_contains(self._column, as_scalar(pat, "str"))
+        else:
+            result_col = cpp_contains_multiple(
+                self._column, column.as_column(pat, dtype="str")
+            )
+        return self._return_or_inplace(result_col, **kwargs)
 
     def replace(
         self, pat, repl, n=-1, case=None, flags=0, regex=True, **kwargs