+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - PR #6460 Add is_timestamp format check API
 - PR #6528 Enable `fixed_point` binary operations
 - PR #6581 Add JNI API to check if PTDS is enabled
+- PR #6598 Add strings::contains API with target column parameter
 
 ## Improvements
 - PR #6430 Add struct type support to `to_arrow` and `from_arrow`